Change reference to bumpedAt to last_posted_at
This addresses a bug where [model.bumpedAt] reports the date of the
thread OP rather than the last bump date. While this will be slightly
less accurate (edits to the last post in a thread will no longer reset
the warning timer), it's way better to remove the spurious warnings.